// src/types/events.ts
export type FactorKey = 'health' | 'spiritual' | 'intelligence' | 'ai' | 'emotion' | 'career' | 'finance' | 'culture' | 'community' | 'environment'

export interface IScore {
  health: number
  spiritual: number
  intelligence: number
  ai: number
  emotion: number
  career: number
  finance: number
  culture: number
  community: number
  environment: number
}